Rename skimage2/feature/corner.py to private _corner.py
