+++ src/_skimage2/feature/__init__.pyi
@@ -57,7 +57,7 @@ from .texture import (
     draw_multiblock_lbp,
 )
 from ._peaks import peak_local_max
-from .corner import (
+from ._corner import (
     corner_kitchen_rosenfeld,
     corner_harris,
     corner_shi_tomasi,
