+++ src/_skimage2/feature/_corner.py
