+++ src/_skimage2/feature/censure.py
@@ -2,7 +2,7 @@ import numpy as np
 from scipy.ndimage import maximum_filter, minimum_filter, convolve
 
 from ..transform import integral_image
-from .corner import structure_tensor
+from ._corner import structure_tensor
 from ..morphology import octagon, star
 from .censure_cy import _censure_dob_loop
 from ..feature.util import (
