+++ src/_skimage2/feature/meson.build
@@ -57,13 +57,13 @@ python_sources = [
   '__init__.pyi',
   '_basic_features.py',
   '_canny.py',
+  '_corner.py',
   '_daisy.py',
   '_hog.py',
   '_orb_descriptor_positions.py',
   'blob.py',
   'brief.py',
   'censure.py',
-  'corner.py',
   '_fisher_vector.py',
   'haar.py',
   'match.py',
